refactor(mobile): use shared GitHub Project sorting
